pyproject.toml
[project]
name = "apibridge-pro"
version = "0.1.0"
description = "Global universal API connector with smart routing, transforms, budgets, and privacy."
requires-python = ">=3.11"
dependencies = [
  "fastapi",
  "uvicorn[standard]",
  "httpx[http2]",
  "orjson",
  "pyyaml",
  "pydantic>=2",
  "redis>=5",
  "jmespath",
  "cryptography>=41.0.0",
  "prometheus-client>=0.19.0",
  "opentelemetry-api>=1.21.0",
  "opentelemetry-sdk>=1.21.0",
  "opentelemetry-exporter-otlp>=1.21.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4.0",
  "pytest-asyncio>=0.21.0",
  "pytest-cov>=4.1.0",
  "respx>=0.20.0",
  "ruff>=0.1.0",
  "mypy>=1.7.0",
  "bandit>=1.7.5",
  "types-PyYAML",
  "types-redis"
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.uvicorn]
reload = true
host = "0.0.0.0"
port = 8000

[tool.ruff]
target-version = "py311"
line-length = 120
select = [
  "E",   # pycodestyle errors
  "W",   # pycodestyle warnings
  "F",   # pyflakes
  "I",   # isort
  "B",   # flake8-bugbear
  "C4",  # flake8-comprehensions
  "UP",  # pyupgrade
  "ARG", # flake8-unused-arguments
  "SIM", # flake8-simplify
]
ignore = [
  "E501",  # line too long (handled by formatter)
  "B008",  # do not perform function calls in argument defaults
  "W191",  # indentation contains tabs
]

[tool.ruff.per-file-ignores]
"tests/*" = ["ARG", "S101"]  # Allow unused args and asserts in tests

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --cov=app --cov-report=term-missing"
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.bandit]
exclude_dirs = ["tests", ".venv"]
skips = ["B101"]  # Allow assert statements

[tool.coverage.run]
source = ["app"]
omit = ["tests/*", "*/test_*.py"]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false

